YOUTUBE/requirements.txt
crewai
crewai_tools
langchain-groq
python-dotenv
crewai[tools]
transformers
